+++ protobuf/lib/src/protobuf/extension_field_set.dart
@@ -164,7 +164,7 @@ class _ExtensionFieldSet {
       final value = original._getFieldOrNull(extension);
       if (value == null) continue;
       if (extension.isRepeated) {
-        assert(value is PbListBase);
+        assert(value is PbList);
         _ensureRepeatedField(extension).addAll(value);
       } else {
         _setFieldUnchecked(extension, value);
@@ -179,12 +179,7 @@ class _ExtensionFieldSet {
       if (field.isRepeated) {
         final entries = _values[field.tagNumber];
         if (entries == null) continue;
-        if (field.isGroupOrMessage) {
-          for (var subMessage in entries as List<GeneratedMessage>) {
-            subMessage.freeze();
-          }
-        }
-        _values[field.tagNumber] = entries.toFrozenPbList();
+        entries.freeze();
       } else if (field.isGroupOrMessage) {
         final entry = _values[field.tagNumber];
         if (entry != null) {