+++ protobuf/lib/src/protobuf/proto3_json.dart
@@ -98,7 +98,7 @@ Object? _writeToProto3Json(_FieldSet fs, TypeRegistry typeRegistry) {
             valueToProto3Json(entryValue, mapEntryInfo.valueFieldType));
       });
     } else if (fieldInfo.isRepeated) {
-      jsonValue = (value as PbListBase)
+      jsonValue = (value as PbList)
           .map((element) => valueToProto3Json(element, fieldInfo.type))
           .toList();
     } else {